Auto-generated utility functions

For each newly defined diffusion DiffusionProcess the package auto-generates some utility functions. They are not exported, but can be called by accessing through DiffusionDefinition module. These are:

Base.eltype — Method

Base.eltype(::DiffusionProcess{T}) where T = T

Return the datatype that each coordinate of the stochastic process is stored in.

Base.zero — Function

Base.zero(K::Type, D, ::Val)

If K is a mutable type, then create zeros of dimension D and entries with types eltype(K). Otherwise, calls regular zero(K).

Base.zero — Method

Base.zero(P::DiffusionProcess)

Instantiate a zero element that can represent a state of a Brownian motion driving a diffusion process.

Base.zero — Method

Base.zero(P::DiffusionProcess)

Instantiate a zero element that can represent a state of a diffusion.

DiffusionDefinition.clone — Method

clone(
    P::T, ξ, glob_to_loc, θ°idx, ::Val{:associate_by_name}
) where T <: DiffusionProcess

Clone diffusion law P substituting relevant parameters with new values. ξ is a vector with parameters (possibly more than needed). ξ itself can be a subvector of a larger vector ξ°. The coordinates relevant for P are given in θ°idx and they refer to positions in the larger vector ξ°. The connection between coordinates in ξ° and ξ is given in glob_to_loc.

DiffusionDefinition.const_parameter_names — Method

const_parameter_names(P::DiffusionProcess)

Return a tuple with the names of all paremeters that are considered to be constant.

DiffusionDefinition.const_parameter_names — Method

const_parameter_names(P::Type{<:DiffusionProcess})

Return a tuple with the names of all paremeters that are considered to be constant.

DiffusionDefinition.const_parameters — Method

const_parameters(P::DiffusionProcess)

Return a tuple of pairs of parameter_name => parameter_value. Return only those parameteres that are considered to be constant.

DiffusionDefinition.default_type — Method

default_type(::DiffusionProcess{T,DP})

Allows for inference of data type that encodes the state space of a given diffusion.

DiffusionDefinition.default_wiener_type — Method

default_wiener_type(::DiffusionProcess{T,DP,DW})

Allows for inference of data type that encodes the state space of the Brownian motion driving a given diffusion process.

DiffusionDefinition.diagonalBmat — Method

diagonalBmat(P::DiffusionProcess)

Indicator for whether the B matrix (if exists) is represented by a diagonal matrix

DiffusionDefinition.diagonaldiff — Method

diagonaldiff(P::DiffusionProcess)

Indicator for whether the volatility coefficient is represented by a diagonal matrix

DiffusionDefinition.dimension — Method

dimension(::DiffusionProcess{T,DP,DW})

Return dimension of the stochastic process and driving Brownian motion.

DiffusionDefinition.end_point_info — Method

end_point_info(P::DiffusionProcess)

Return information about the end-point (works only if some information of this kind has been passed at the time of defining a struct) TODO improve

DiffusionDefinition.end_point_info_names — Method

end_point_info_names(P::DiffusionProcess)

Return names of information pieces about the end-point (works only if some information of this kind has been passed at the time of defining a struct) TODO improve

DiffusionDefinition.parameter_names — Method

parameter_names(P::DiffusionProcess)

Return a tuple with the names of all paremeters.

DiffusionDefinition.parameter_names — Method

parameter_names(::Type{<:DiffusionProcess})

Return a tuple with the names of all paremeters.

DiffusionDefinition.parameters — Method

parameters(P::DiffusionProcess)

Return a tuple of pairs of parameter_name => parameter_value.

DiffusionDefinition.sparseBmat — Method

sparseBmat(P::DiffusionProcess)

Indicator for whether the B matrix (if exists) is represented by a sparse matrix

DiffusionDefinition.sparsediff — Method

sparsediff(P::DiffusionProcess)

Indicator for whether the volatility coefficient is represented by a sparse matrix

DiffusionDefinition.state_space — Method

state_space(::DiffusionProcess{T,DP,DW,SS})

Return the state space restrictions.

DiffusionDefinition.var_parameter_names — Method

var_parameter_names(P::DiffusionProcess)

Return a tuple with the names of all paremeters that are considered to be variable.

DiffusionDefinition.var_parameter_names — Method

var_parameter_names(P::Type{<:DiffusionProcess})

Return a tuple with the names of all paremeters that are considered to be variable.

DiffusionDefinition.var_parameters — Method

var_parameters(P::DiffusionProcess)

Return a tuple of pairs of parameter_name => parameter_value. Return only those parameteres that are considered to be variable.