Index

DiffusionDefinition.AbstractBuffer

DiffusionDefinition.AbstractSDESolver

DiffusionDefinition.BoundedStateSpace

DiffusionDefinition.DiffusionProcess

DiffusionDefinition.DiffusionStateSpace

DiffusionDefinition.EulerMaruyama

DiffusionDefinition.LinearDiffBuffer

DiffusionDefinition.LinearDiffusion

DiffusionDefinition.LowerBoundedStateSpace

DiffusionDefinition.StandardEulerBuffer

DiffusionDefinition.UnboundedStateSpace

DiffusionDefinition.UpperBoundedStateSpace

DiffusionDefinition.Wiener

Base.Unicode.lowercase

Base.eltype

Base.rand

Base.zero

Base.zero

Base.zero

DiffusionDefinition.B

DiffusionDefinition.B!

DiffusionDefinition._is_datatype

DiffusionDefinition._symbol_in

DiffusionDefinition._symbol_in

DiffusionDefinition._symbol_in

DiffusionDefinition.a

DiffusionDefinition.a!

DiffusionDefinition.add_diff_function!

DiffusionDefinition.add_nonhypo_function!

DiffusionDefinition.add_phi_function!

DiffusionDefinition.b

DiffusionDefinition.b!

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.cleanup_param_names!

DiffusionDefinition.clone

DiffusionDefinition.const_parameter_names

DiffusionDefinition.const_parameter_names

DiffusionDefinition.const_parameters

DiffusionDefinition.const_parameters

DiffusionDefinition.constdiff

DiffusionDefinition.createstruct

DiffusionDefinition.default_type

DiffusionDefinition.default_type

DiffusionDefinition.default_wiener_type

DiffusionDefinition.default_wiener_type

DiffusionDefinition.diagonalBmat

DiffusionDefinition.diagonaldiff

DiffusionDefinition.dimension

DiffusionDefinition.dimensions

DiffusionDefinition.end_point_info

DiffusionDefinition.end_point_info_names

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.get_curly

DiffusionDefinition.get_name_stem

DiffusionDefinition.highest_idx_used

DiffusionDefinition.parameter_names

DiffusionDefinition.parameter_names

DiffusionDefinition.parameters

DiffusionDefinition.parameters

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_lines!

DiffusionDefinition.parse_param_multi_names

DiffusionDefinition.parse_param_single_name

DiffusionDefinition.parse_process

DiffusionDefinition.prepare_abstract_type

DiffusionDefinition.process_name

DiffusionDefinition.remove_curly

DiffusionDefinition.solve!

DiffusionDefinition.solve!

DiffusionDefinition.sparseBmat

DiffusionDefinition.sparsediff

DiffusionDefinition.state_space

DiffusionDefinition.update_label

DiffusionDefinition.var_parameter_names

DiffusionDefinition.var_parameter_names

DiffusionDefinition.var_parameters

DiffusionDefinition.var_parameters

DiffusionDefinition.β

DiffusionDefinition.β!

DiffusionDefinition.σ

DiffusionDefinition.σ!

Random.rand!

Random.rand!

Random.rand!

Trajectories.trajectory

Trajectories.trajectory

DiffusionDefinition.@diffusion_process

DiffusionDefinition.@load_diffusion

DiffusionDefinition.@load_diffusion

DiffusionDefinition.AbstractBuffer — Type

AbstractBuffer{T} <: AbstractArray{T,1}

Types inheriting from AbstractBuffer define buffers used for various in-place computations. Each subtype MUST:

have a field data which is a one-dimensional vector with data

provide a constructor of the form: NAME{eltype}(dim1, dim2, ...)

and it CAN:

have multiple fields that provide views or reshaped views to data

DiffusionDefinition.BoundedStateSpace — Type

BoundedStateSpace{L,U} <: DiffusionStateSpace

Upper and lower bounds imposed on the state-space of a diffusion process. L corresponds to lower bounds, U corresponds to upper bounds.

DiffusionDefinition.LowerBoundedStateSpace — Type

LowerBoundedStateSpace{T,S,N} <: DiffusionStateSpace

Lower bounds imposed on the state-space of a diffusion process. T is used to list the indices that have lower-bound restrictions, S indicates the values of the lower-bounds, N is the total number of coordinates with lower-bound restrictions

DiffusionDefinition.UnboundedStateSpace — Type

UnboundedStateSpace <: DiffusionStateSpace

No restrictions imposed on the state-space of the process (i.e. ℝᵈ)

DiffusionDefinition.UpperBoundedStateSpace — Type

UpperBoundedStateSpace{T,S,N} <: DiffusionStateSpace

Upper bounds imposed on the state-space of a diffusion process. T is used to list the indices that have upper-bound restrictions, S indicates the values of the upper-bounds, N is the total number of coordinates with upper-bound restrictions

DiffusionDefinition.Wiener — Type

struct Wiener{D,Tdevice}
end

A struct defining the Wiener process. D indicates the dimension of the process if it cannot be inferred from the DataType of the Trajectory. If the dimension can be inferred then it takes precedence over the value of D.

Trajectories.trajectory — Method

Trajectories.trajectory(tt, v::Type, D::Number)

Create a Trajectory with mutable states of dimension D along the time collection tt.

Trajectories.trajectory — Method

Trajectories.trajectory(tt, v::Type)

Create a Trajectory with immutable states along the time collection tt.

DiffusionDefinition.@diffusion_process — Macro

diffusion_process(name, ex::Expr, p...)

Defines a diffusion process according to a template described in the documentation of the github repository: https://github.com/mmider/DiffusionDefinition.jl

DiffusionDefinition.@load_diffusion — Macro

load_diffusion(name)

Loads the predefined diffusion process.

DiffusionDefinition.@load_diffusion — Macro

load_diffusion()

Displays available choices of predefined diffusion processes that can be loaded

DiffusionDefinition.AbstractSDESolver — Type

abstract type AbstractSDESolver end

Supertype of flags indicating ODE solvers

DiffusionDefinition.DiffusionProcess — Type

DiffusionProcess{T,DP,DW,SS,EI}

Types inheriting from DiffusionProcess define Ito diffusions. T denotes the datatype of each coordinate, DP the dimension of the stochastic process, DW the dimension of the Wiener process, SS lists the state space restrictions

DiffusionDefinition.DiffusionStateSpace — Type

DiffusionDomain

Types inheriting from DiffusionStateSpace define the types of restrictions put on the state-space of the stochastic process.

DiffusionDefinition.EulerMaruyama — Type

struct EulerMaruyama <: AbstractSDESolver end

Flag for indicating use of the Euler-Maruyama scheme for sampling diffusions. IMPORTANT: this is the only diffusion path sampler implemented in this package. There are no plans for implementing any other SDE solver in the forseeable future.

DiffusionDefinition.LinearDiffBuffer — Type

struct LinearDiffBuffer{T,D,Tb,TB,Tσ,Tdw} <: AbstractBuffer{T}
    data::Vector{T}
    b::Tb
    B::TB
    σ::Tσ
    dW::Tdw
end

A buffer for Euler-Maruyama simulations of linear diffusions. Almost the same as StandardEulerBuffer, but contains additional space for an intermediate construction of a matrix B (and a corrsponding view).

DiffusionDefinition.LinearDiffusion — Type

LinearDiffusion{T,DP,DW,SS} <: DiffusionProcess{T,DP,DW,SS}

Types inheriting from LinearDiffusion define a linear Ito-type diffusion, i.e. solutions to stochastic differential equations of the form: dXₜ = (BₜXₜ + βₜ)dt + σₜdWₜ, t∈[0,T], X₀=x₀.

DiffusionDefinition.StandardEulerBuffer — Type

struct StandardEulerBuffer{T,D,Tb,Tσ,Tdw} <: AbstractBuffer{T}
    data::Vector{T}
    b::Tb
    σ::Tσ
    dW::Tdw
end

Standard buffer for the Euler-Maruyama simulations. The data is stored in data and b, σ, dW provide appropriately reshaped views to the corresponding segments of data. T is the DataType of each data element and D are dimensions needed to create the views.

Base.Unicode.lowercase — Method

Base.lowercase(s::Symbol)

Lowercase all letters in a symbol

Base.rand — Function

Base.rand(w::Wiener, tt, y1)

Samples Wiener process on tt, started from y1 and returns a new object with a sampled trajectory.

DiffusionDefinition.B — Function

B

Compute matrix B of a linear diffusion out-of-place (should use StaticArrays).

DiffusionDefinition.B! — Function

B!

Compute matrix B of a linear diffusion in-place (uses buffers to store temporary results).

DiffusionDefinition._is_datatype — Method

_is_datatype(sym, p)

Utility function that checks whether sym is a datatype. It returns true if sym is either an in-built datatype (say float, StaticArray etc.) or if it is a template argument.

DiffusionDefinition._symbol_in — Method

_symbol_in(::Any, ::Any)

Return false by default i.e. if the first argument is not a symbol nor its quote

DiffusionDefinition._symbol_in — Method

_symbol_in(s::QuoteNode, symbols)

Check if quote of a symbol s is listed in a list of symbols symbols

DiffusionDefinition._symbol_in — Method

_symbol_in(s::Symbol, symbols)

Check if symbol s is listed in a list of symbols symbols

DiffusionDefinition.a — Function

a

Compute the diffusion function σσ' of a diffusion out-of-place (should use StaticArrays).

DiffusionDefinition.a! — Function

a!

Compute the diffusion function σσ' of a diffusion in-place (uses buffers to store temporary results).

DiffusionDefinition.add_diff_function! — Method

add_diff_function!(fns, p)

Add a definition of a function consdiff that indicates if the diffusion coefficient is constant

DiffusionDefinition.add_nonhypo_function! — Method

add_nonhypo_function!(fns, line, p)

Add a definition of a function nonhypo that takes a current diffusion state x and returns the elliptic coordinates.

DiffusionDefinition.add_phi_function! — Method

add_phi_function!(expr::Expr, fns::Vector{Expr}, t::Symbol, x::Symbol, p)

Construct a phi function used for conjugate updates.

DiffusionDefinition.b — Function

b

Compute the drift of a diffusion out-of-place (should use StaticArrays).

DiffusionDefinition.b! — Function

b!

Compute the drift of a diffusion in-place (uses buffers to store temporary results).

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::UnboundedStateSpace, x)

No restrictions, bounds satisfied by default

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::LowerBoundedStateSpace{T,S,N}, x) where {T,S,N}

Checks if all coordinates adhere to lower bound restrictions

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::UpperBoundedStateSpace{T,S,N}, x) where {T,S,N}

Checks if all coordinates adhere to upper bound restrictions

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::BoundedStateSpace{L,U}, x) where {L,U}

Checks if all coordinates adhere to lower and upper bound restrictions

DiffusionDefinition.cleanup_param_names! — Method

cleanup_param_names!(expr, t::Symbol, x::Symbol, params)

A hepler function for cleaning up the names of variables in the definition of a phi function.

DiffusionDefinition.constdiff — Function

constdiff

Returns true if the diffusion coefficient does not depend on the state of the process or time.

DiffusionDefinition.createstruct — Method

createstruct(abstract_type, name, params)

Create code that defines a struct defining a diffusion process.

DiffusionDefinition.dimensions — Method

dimensions(::Type{K}) where K<:_AB

Utility function for structs inheriting from AbstractBuffer. Returns the dimensions required for constructing the internal views of the data.

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(p)

Fill all unspecified variables with default values

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(::Val{:additional}, p)

If unspecified, there are no restriction on a state space, the volatility coefficient is assumed constant, the diffusion is not linear and the datatype of each coordinate is set to Float64.

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(::Val{:dimensions}, p)

If unspecified, the dimension of the stochastic process and the driving Brownian motion is set to 1.

DiffusionDefinition.get_curly — Method

get_curly(::Type{K}) where K

Utility function that returns a tuple with all type-specifiers listed in the curly brackets.

Examples

julia> remove_curly(Array{Float64,1})
(:Float64, 1)

DiffusionDefinition.get_name_stem — Method

get_name_stem(name_stem::Symbol, parameters)

Get the stem of a name for a parameters and then add a disambiguation index. Underscore _ used in place of name is defaulted to p.

DiffusionDefinition.highest_idx_used — Method

highest_idx_used(name_stem, params)

Find the highest disambiguation index that has been used thus far for a given name_stem.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:additional}, line, p)

Parse a line that defines additional information about a diffusion process.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:aux_info}, line, p)

Parse a line that defines parameters of the diffusion. The line must be in a format: name –> parameter-description

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:conjugate}, line, p)

Parse a line that defines conjugate updates.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:dimensions}, line, p)

Parse a line that defines the dimension of a diffusion process and the driving Brownian motion.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:parameters}, line, p)

Parse a line that defines parameters of the diffusion. The line must be in a format: name –> parameter-description

DiffusionDefinition.parse_lines! — Method

parse_lines!(ex::Expr, p, condition)

Parse all lines of the expression ex, but process only those which satisfy condition. p is a passed-around structure that accumulates processed information.

DiffusionDefinition.parse_param_multi_names — Function

parse_param_multi_names(line, p)

Parse a line that defines parameters of the diffusion. The line must be in one of the formats: (pname1, pname2, ...) –> (numberofparameters, datatype) (pname1, pname2, ...) –> datatype (pname1, pname2, ...) –> (datatype1, datatype2, ...) In the former two cases defines number_of_parameters-many parameters, with names p_name1, p_name2, ... and of datatype type. In the last case the datatypes differ from parameter to parameter.

DiffusionDefinition.parse_param_single_name — Function

parse_param_single_name(line, p)

Parse a line that defines parameters of the diffusion. The line must be in one of the formats: parametername –> (numberofparameters, datatype) parametername –> datatype In the former case defines number_of_parameters-many parameters, with names parameter_namei and of datatype type. In the latter case defines a single parameters with name parameter_name and of type datatype.

DiffusionDefinition.parse_process — Method

parse_process(name , ex::Expr, ::Any)

Parse a template defining a diffusion process, create a corresponding struct and specified functions, evaluate them in the environment of a package and then import the struct name to Main scope, in which the package has been imported to.

DiffusionDefinition.prepare_abstract_type — Method

prepare_abstract_type(stem, dims, data_type, state_restr)

Create a string defining a parent, abstract type from its stem, the dimensions dims of the process and the driving Brownian motion, the datatype data_type of each coordinate and the restrictions on the state space state_restr.

DiffusionDefinition.process_name — Function

process_name

Process the name of a struct by returning the pure name (without template arguments), the name with template arguments and a list of template arguments.

DiffusionDefinition.remove_curly — Method

remove_curly(::Type{K}) where K

Utility function that removes all type-specifiers listed in the curly brackets.

Examples

julia> remove_curly(Array{Float64,1})
Array

DiffusionDefinition.update_label — Method

update_label(line, current_label)

Update the label, which signifies what type of information a given line in a template is supposed to be encoding.

DiffusionDefinition.β — Function

β

Compute vector β of a linear diffusion out-of-place (should use StaticArrays).

DiffusionDefinition.β! — Function

β!

Compute vector β! of a linear diffusion in-place (uses buffers to store temporary results).

DiffusionDefinition.σ — Function

σ

Compute the volatility coefficient of a diffusion out-of-place (should use StaticArrays).

DiffusionDefinition.σ! — Function

σ!

Compute the volatility coefficient of a diffusion in-place (uses buffers to store temporary results).

Random.rand! — Method

Random.rand!(
    w::Wiener{D},
    path::Trajectory{T,Vector{K}},
    y1=zero(K,D,ismutable(K))
) where {T,K,D}

Samples Wiener process started from y1 and saves the data in path. Uses a default random number generator.

Random.rand! — Method

Random.rand!(
    rng::Random.AbstractRNG,
    ::Wiener{D},
    path::Trajectory{T,Vector{Vector{K}}},
    y1=zeros(K,D)
) where {K,T,D}

Samples Wiener process with mutable states, started from y1 and saves the data in path. rng is used as a random number generator.

Random.rand! — Method

Random.rand!(
    rng::Random.AbstractRNG,
    ::Wiener,
    path::Trajectory{T,Vector{K}},
    y1=zero(K)
) where {T,K}

Samples Wiener process with immutable states, started from y1 and saves the data in path. rng is used as a random number generator.