DiffusionDefinition.jl

This is a small utility package for defining diffusion processes, created to work in conjunction with BridgeSDEInference.jl.

DiffusionDefinition.BoundedStateSpace

DiffusionDefinition.DiffusionProcess

DiffusionDefinition.DiffusionStateSpace

DiffusionDefinition.LinearDiffusion

DiffusionDefinition.LowerBoundedStateSpace

DiffusionDefinition.UnboundedStateSpace

DiffusionDefinition.UpperBoundedStateSpace

Base.Unicode.lowercase

Base.eltype

DiffusionDefinition._symbol_in

DiffusionDefinition._symbol_in

DiffusionDefinition._symbol_in

DiffusionDefinition.add_constdiff_function!

DiffusionDefinition.add_nonhypo_function!

DiffusionDefinition.add_phi_function!

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.bound_satisfied

DiffusionDefinition.cleanup_param_names!

DiffusionDefinition.createstruct

DiffusionDefinition.dimension

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.fill_unspecified_with_defaults

DiffusionDefinition.get_name_stem

DiffusionDefinition.highest_idx_used

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_line!

DiffusionDefinition.parse_lines!

DiffusionDefinition.parse_param_multi_names

DiffusionDefinition.parse_param_single_name

DiffusionDefinition.parse_process

DiffusionDefinition.prepare_abstract_type

DiffusionDefinition.state_space

DiffusionDefinition.update_label

DiffusionDefinition.@diffusion_process

DiffusionDefinition.@load_diffusion

DiffusionDefinition.@load_diffusion

DiffusionDefinition.BoundedStateSpace — Type

BoundedStateSpace{L,U} <: DiffusionStateSpace

Upper and lower bounds imposed on the state-space of a diffusion process. L corresponds to lower bounds, U corresponds to upper bounds.

DiffusionDefinition.LowerBoundedStateSpace — Type

LowerBoundedStateSpace{T,S,N} <: DiffusionStateSpace

Lower bounds imposed on the state-space of a diffusion process. T is used to list the indices that have lower-bound restrictions, S indicates the values of the lower-bounds, N is the total number of coordinates with lower-bound restrictions

DiffusionDefinition.UnboundedStateSpace — Type

UnboundedStateSpace <: DiffusionStateSpace

No restrictions imposed on the state-space of the process (i.e. ℝᵈ)

DiffusionDefinition.UpperBoundedStateSpace — Type

UpperBoundedStateSpace{T,S,N} <: DiffusionStateSpace

Upper bounds imposed on the state-space of a diffusion process. T is used to list the indices that have upper-bound restrictions, S indicates the values of the upper-bounds, N is the total number of coordinates with upper-bound restrictions

DiffusionDefinition.@diffusion_process — Macro

diffusion_process(name, ex::Expr, p...)

Defines a diffusion process according to a template described in the documentation of the github repository: https://github.com/mmider/DiffusionDefinition.jl

DiffusionDefinition.@load_diffusion — Macro

load_diffusion(name)

Loads the predefined diffusion process.

DiffusionDefinition.@load_diffusion — Macro

load_diffusion()

Displays available choices of predefined diffusion processes that can be loaded

DiffusionDefinition.DiffusionProcess — Type

DiffusionProcess{T,DP,DW}

Types inheriting from DiffusionProcess define Ito diffusions. T denotes the datatype of each coordinate, DP the dimension of the stochastic process, DW the dimension of the Wiener process, SS lists the state space restrictions.

DiffusionDefinition.DiffusionStateSpace — Type

DiffusionDomain

Types inheriting from DiffusionStateSpace define the types of restrictions put on the state-space of the stochastic process.

DiffusionDefinition.LinearDiffusion — Type

LinearDiffusion{T,DP,DW,SS} <: DiffusionProcess{T,DP,DW,SS}

Types inheriting from LinearDiffusion define a linear Ito-type diffusion, i.e. solutions to stochastic differential equations of the form: dXₜ = (BₜXₜ + βₜ)dt + σₜdWₜ, t∈[0,T], X₀=x₀.

Base.Unicode.lowercase — Method

lowercase(s::Symbol)

Lowercase all letters in a symbol

Base.eltype — Method

eltype(::DiffusionProcess{T}) where T = T

Return the datatype that each coordinate of the stochastic process is stored in.

DiffusionDefinition._symbol_in — Method

_symbol_in(::Any, ::Any)

Return false by default i.e. if the first argument is not a symbol nor its quote

DiffusionDefinition._symbol_in — Method

_symbol_in(s::QuoteNode, symbols)

Check if quote of a symbol s is listed in a list of symbols symbols

DiffusionDefinition._symbol_in — Method

_symbol_in(s::Symbol, symbols)

Check if symbol s is listed in a list of symbols symbols

DiffusionDefinition.add_constdiff_function! — Method

add_constdiff_function!(fns, p)

Add a definition of a function consdiff that indicates if the diffusion coefficient is constant

DiffusionDefinition.add_nonhypo_function! — Method

add_nonhypo_function!(fns, line, p)

Add a definition of a function nonhypo that takes a current diffusion state x and returns the elliptic coordinates.

DiffusionDefinition.add_phi_function! — Method

add_phi_function!(expr::Expr, fns::Vector{Expr}, t::Symbol, x::Symbol, p)

Construct a phi function used for conjugate updates.

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::UnboundedStateSpace, x)

No restrictions, bounds satisfied by default

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::LowerBoundedStateSpace{T,S,N}, x) where {T,S,N}

Checks if all coordinates adhere to lower bound restrictions

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::UpperBoundedStateSpace{T,S,N}, x) where {T,S,N}

Checks if all coordinates adhere to upper bound restrictions

DiffusionDefinition.bound_satisfied — Method

bound_satisfied(::BoundedStateSpace{L,U}, x) where {L,U}

Checks if all coordinates adhere to lower and upper bound restrictions

DiffusionDefinition.cleanup_param_names! — Method

cleanup_param_names!(expr, t::Symbol, x::Symbol, params)

A hepler function for cleaning up the names of variables in the definition of a phi function.

DiffusionDefinition.createstruct — Method

createstruct(abstract_type, name, params)

Create code that defines a struct defining a diffusion process.

DiffusionDefinition.dimension — Method

dimension(::DiffusionProcess{T,DP,DW})

Return dimension of the stochastic process and driving Brownian motion.

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(p)

Fill all unspecified variables with default values

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(::Val{:additional}, p)

If unspecified, there are no restriction on a state space, the volatility coefficient is assumed constant, the diffusion is not linear and the datatype of each coordinate is set to Float64.

DiffusionDefinition.fill_unspecified_with_defaults — Method

fill_unspecified_with_defaults(::Val{:dimensions}, p)

If unspecified, the dimension of the stochastic process and the driving Brownian motion is set to 1.

DiffusionDefinition.get_name_stem — Method

get_name_stem(name_stem::Symbol, parameters)

Get the stem of a name for a parameters and then add a disambiguation index. Underscore _ used in place of name is defaulted to p.

DiffusionDefinition.highest_idx_used — Method

highest_idx_used(name_stem, params)

Find the highest disambiguation index that has been used thus far for a given name_stem.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:additional}, line, p)

Parse a line that defines additional information about a diffusion process.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:conjugate}, line, p)

Parse a line that defines conjugate updates.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:dimensions}, line, p)

Parse a line that defines the dimension of a diffusion process and the driving Brownian motion.

DiffusionDefinition.parse_line! — Method

parse_line!(::Val{:parameters}, line, p)

Parse a line that defines parameters of the diffusion. The line must be in a format: name –> parameter-description

DiffusionDefinition.parse_lines! — Method

parse_lines!(ex::Expr, p, condition)

Parse all lines of the expression ex, but process only those which satisfy condition. p is a passed-around structure that accumulates processed information.

DiffusionDefinition.parse_param_multi_names — Method

parse_param_multi_names(line, p)

Parse a line that defines parameters of the diffusion. The line must be in one of the formats: (pname1, pname2, ...) –> (numberofparameters, datatype) (pname1, pname2, ...) –> datatype (pname1, pname2, ...) –> (datatype1, datatype2, ...) In the former two cases defines number_of_parameters-many parameters, with names p_name1, p_name2, ... and of datatype type. In the last case the datatypes differ from parameter to parameter.

DiffusionDefinition.parse_param_single_name — Method

parse_param_single_name(line, p)

Parse a line that defines parameters of the diffusion. The line must be in one of the formats: parametername –> (numberofparameters, datatype) parametername –> datatype In the former case defines number_of_parameters-many parameters, with names parameter_namei and of datatype type. In the latter case defines a single parameters with name parameter_name and of type datatype.

DiffusionDefinition.parse_process — Method

parse_process(name , ex::Expr, ::Any)

Parse a template defining a diffusion process, create a corresponding struct and specified functions, evaluate them in the environment of a package and then import the struct name to Main scope, in which the package has been imported to.

DiffusionDefinition.prepare_abstract_type — Method

prepare_abstract_type(stem, dims, data_type, state_restr)

Create a string defining a parent, abstract type from its stem, the dimensions dims of the process and the driving Brownian motion, the datatype data_type of each coordinate and the restrictions on the state space state_restr.

DiffusionDefinition.state_space — Method

state_space(::DiffusionProcess{T,DP,DW,SS})

Return the state space restrictions.

DiffusionDefinition.update_label — Method

update_label(line, current_label)

Update the label, which signifies what type of information a given line in a template is supposed to be encoding.